Mental Health Management Ontology - Latest Version
This ontology can be used to assist in the management of mental healthcare network to allow the integration and interoperability between independent databases and to enable the generation of relevant indicators to mental health, to improve decision making, increasing the efficiency and effectiveness of mental health management. Ontology created in the Laboratory of Health Intelligence - Ribeirao Preto Medical School, University of Sao Paulo, Brazil.